Update props in RocketsPage.jsx
