+++ src/App.js
@@ -1,11 +1,19 @@
+import React, { useEffect } from 'react';
 import { Routes, Route } from 'react-router-dom';
-import Rockets from './pages/RocketsPage';
+import { useDispatch } from 'react-redux';
+import Rockets from './pages/rockets/RocketsPage';
 import Missions from './pages/MissionsPage';
 import Profile from './pages/ProfilePage';
 import Layout from './components/layout/Layout';
 import './App.css';
+import { fetchRocketsData } from './redux/rockets/rocketsSlice';
 
 function App() {
+  const dispatch = useDispatch();
+
+  useEffect(() => {
+    dispatch(fetchRocketsData());
+  }, [dispatch]);
   return (
     <div>
       <Layout>
