--- src/pages/RocketsPage.jsx
@@ -1,7 +1,0 @@
-import React from 'react';
-
-const Rockets = () => (
-  <div>Rockets</div>
-);
-
-export default Rockets;
