+++ src/pages/rockets/Rockets.module.css
@@ -1,0 +1,5 @@
+.rocket__list {
+  display: flex;
+  flex-direction: column;
+  gap: 30px;
+}
